using tfIdf matrix instead of search query

+++ IR-Exercise4/src/ClusteringAlgorithms/Centroid.java
@@ -5,19 +5,19 @@ package ClusteringAlgorithms;
  */
 public class Centroid implements ICentroid
 {
-    private double coordinates[];
+    private float coordinates[];
 
-    public Centroid(double[] coordinates)
+    public Centroid(float[] coordinates)
     {
         this.coordinates = coordinates;
     }
 
     public Centroid(int numberOfDimensions)
     {
-        coordinates = new double[numberOfDimensions];
+        coordinates = new float[numberOfDimensions];
     }
 
-    public double []GetCoordinates()
+    public float[] GetCoordinates()
     {
         return coordinates;
     }
