+++ IR-Exercise4/src/ClusteringAlgorithms/ICentroid.java
@@ -5,5 +5,5 @@ package ClusteringAlgorithms;
  */
 public interface ICentroid
 {
-    double []GetCoordinates();
+    float[] GetCoordinates();
 }
